Refactor metadata_service
