--- metadata_service/routers/__init__.py
@@ -1,2 +1,0 @@
-DB_NAME = "metadata"
-MONGODB_URL = "mongodb://localhost:27017"
