+++ metadata_service/routes/__init__.py
@@ -13,6 +13,5 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Metadata Service API"""
-
-__version__ = "0.1.0"
+DB_NAME = "metadata"
+MONGODB_URL = "mongodb://localhost:27017"
