+++ metadata_service/routes/experiments.py
@@ -1,10 +1,25 @@
+# Copyright 2021 Universität Tübingen, DKFZ and EMBL
+# for the German Human Genome-Phenome Archive (GHGA)
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
 from typing import List
 
 import motor.motor_asyncio
 from fastapi import APIRouter, HTTPException
 
-from metadata_service.schemas.experiment import Experiment
-from metadata_service.routers import DB_NAME, MONGODB_URL
+from metadata_service.models import Experiment
+from metadata_service.routes import DB_NAME, MONGODB_URL
 
 EXPERIMENT_COLLECTION = "experiments"
