--- metadata_service/schemas/publication.py
@@ -1,8 +1,0 @@
-from typing import Optional
-
-from pydantic import BaseModel
-
-
-class Publication(BaseModel):
-    id: str
-    title: Optional[str] = None
